Creation of the repository - 04/05/2022
To run the code, simply run the main program. It will scan through folders and preserve data using SimpleNamespace variables. Features: 
--> Flight envelope diagrams (Manoeuvring, Gust, Combined)
--> Balancing Loads (Main wing, Horizontal tailplane)
--> TO BE IMPLEMENTED: Shear, Bending and Torsion distribution along the main wing semi span; in general, complete coverage of the CS-VLA, Subpart C.
Soon will be available a sub-routine to dump data to a conveniently formatted file. This code is an example and must be used with care (IT IS NOT YET CHECKED FOR DESIGN PURPOSES).